Separated comments functionality into CommentsService
- Separated Comments and Subject operations
- Moved Controller annotations into controller classes only and simplified ServiceApi interfaces
- Fixed bug with isFlagged being null when creating a new comment
